requirements.txt
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
discord==2.3.2
discord.py==2.4.0
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyNaCl==1.5.0
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
spotipy==2.24.0
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.17.1
yt-dlp==2024.11.4